# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Food FiApp
  version: 1.0.0
  description: API para gerenciar usuarios (dono de restaurante e cliente), restaurantes e itens de cardapio.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /users/{userId}/addresses:
    parameters:
      - name: userId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID unico do usuario.


    get:
      tags:
        - addresses
      security:
        - bearerAuth: []
      operationId: listAddressesByUserId
      summary: Lista todos os enderecos de um usuario.
      description: Retorna uma lista de todos os enderecos associados a um usuario especifico.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressesResponse'
          description: Lista com sucesso todos os enderecos do usuario. Pode retornar uma lista vazia.

    post:
      tags:
        - addresses
      security:
        - bearerAuth: []
      operationId: createAddressesForUser
      summary: Cria um novo endereco para um usuario.
      description: Adiciona um novo endereco a lista de enderecos de um usuario existente.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddressesRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressesResponse'
          description: Cria com sucesso um endereco para o usuario. Retorna os detalhes do endereco recem-criado.
  /users/{userId}/addresses/{addressesId}:
    parameters:
      - name: userId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID unico do usuario proprietario do endereco.
      - name: addressesId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID unico do endereco.


    put:
      tags:
        - addresses
      security:
        - bearerAuth: []
      operationId: updateAddressesForUser
      summary: Atualiza um endereco especifico de um usuario.
      description: Permite a atualizacao dos detalhes de um endereco especifico associado a um usuario.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressesRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressesResponse'
          description: Endereco atualizado com sucesso. Retorna os detalhes atualizados do endereco.
        '404':
          description: Usuario ou endereco nao encontrado.

    delete:
      tags:
        - addresses
      security:
        - bearerAuth: []
      operationId: deleteAddressesForUser
      summary: Deleta logicamente um endereco de um usuario.
      description: Marca um endereco especifico como inativo para um determinado usuario, nao o removendo fisicamente do banco de dados.
      responses:
        '204':
          description: Endereco deletado com sucesso. Nenhuma resposta de conteudo.
        '404':
          description: Usuario ou endereco nao encontrado.

  # Rotas de restaurantes - todas aninhadas sob usuário para validação
...
  /users/{userId}/restaurants/{restaurantId}/menu-items:
    parameters:
      - name: userId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID único do usuário proprietário do restaurante.
      - name: restaurantId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID único do restaurante.

    get:
      operationId: listMenuItems
      tags:
        - menu-items
      summary: Lista os itens do cardápio de um restaurante
      description: Retorna todos os itens do cardápio de um restaurante específico.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de itens do cardápio retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemResponse'

    post:
      operationId: createMenuItem
      tags:
        - menu-items
      summary: Adiciona um novo item ao cardápio
      description: Adiciona um novo item ao cardápio de um restaurante. Apenas o proprietário do restaurante pode realizar esta operação.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome do item do cardápio
                description:
                  type: string
                  description: Descrição do item
                price:
                  type: number
                  format: double
                  description: Preço do item
                availableForInStoreOnly:
                  type: boolean
                  description: Disponível apenas para consumo no local
                photo:
                  type: string
                  format: binary
                  description: Arquivo de imagem do item
      responses:
        '201':
          description: Item do cardápio criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
  /users/{userId}/restaurants/{restaurantId}/menu-items/{itemId}:
    parameters:
      - name: userId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID único do usuário proprietário do restaurante.
      - name: restaurantId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID único do restaurante.
      - name: itemId
        schema:
          format: uuid
          type: string
        in: path
        required: true
        description: O ID único do item do cardápio.

    get:
      operationId: getMenuItem
      tags:
        - menu-items
      summary: Busca um item do cardápio pelo ID
      description: Retorna os detalhes de um item específico do cardápio.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Item do cardápio encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '404':
          description: Item do cardápio não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateMenuItem
      tags:
        - menu-items
      summary: Atualiza um item do cardápio
      description: Atualiza os dados de um item do cardápio. Apenas o proprietário do restaurante pode realizar esta operação.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome do item do cardápio
                description:
                  type: string
                  description: Descrição do item
                price:
                  type: number
                  format: double
                  description: Preço do item
                availableForInStoreOnly:
                  type: boolean
                  description: Disponível apenas para consumo no local
                photo:
                  type: string
                  format: binary
                  description: Arquivo de imagem do item
      responses:
        '200':
          description: Item do cardápio atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '403':
          description: Acesso negado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Item do cardápio não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteMenuItem
      tags:
        - menu-items
      summary: Remove um item do cardápio
      description: Remove um item do cardápio. Apenas o proprietário do restaurante pode realizar esta operação.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Item do cardápio removido com sucesso.
        '403':
          description: Acesso negado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Item do cardápio não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Insira o token JWT obtido do endpoint /auth/login no formato: Bearer {token}
  schemas:
    ErrorResponse:
      title: Resposta de Erro
      required:
        - message
        - code
        - status
        - path
        - timestamp
      type: object
      properties:
        message:
          description: Mensagem de erro detalhada.
          type: string
          example: Dados de entrada invalidos.
        code:
          description: Codigo interno do erro.
          type: string
          example: INVALID_INPUT
        status:
          description: Codigo de status HTTP.
          type: integer
          example: 400
        path:
          description: URL da requisicao que gerou o erro.
          type: string
          example: /api/v1/users
        timestamp:
          description: Data e hora quando o erro ocorreu.
          type: string
          format: date-time
          example: '2025-09-23T22:30:00-03:00'
        fieldErrors:
          description: Mapa de erros de validacao especificos do campo, onde a chave e o nome do campo e o valor e a mensagem de erro.
          type: object
          additionalProperties:
            type: string
          example:
            email: O formato do e-mail e invalido.
            password: A senha deve ter no minimo 8 caracteres.
...
    CreateAddressesRequest:
      title: Solicitacao de Criacao de Endereco
      description: Estrutura da requisicao para criar um novo endereco para um usuario.
      type: object
      properties:
        publicPlace:
          type: string
          description: Nome da rua, avenida, praca, etc.
          example: Rua das Flores
        number:
          type: string
          description: Numero do imovel.
          example: '123'
        complement:
          type: string
          description: Complemento do endereco (apartamento, bloco, etc.).
          example: Apto 45
        neighborhood:
          type: string
          description: Nome do bairro.
          example: Jardim Primavera
        city:
          type: string
          description: Nome da cidade.
          example: Sao Paulo
        state:
          type: string
          description: Sigla do estado (UF).
          example: SP
        postalCode:
          type: string
          description: Codigo de Enderecamento Postal (CEP).
          example: 01234-567
    UpdateAddressesRequest:
      title: Solicitacao de Atualizacao de Endereco
      description: Estrutura da requisicao para atualizar um endereco existente de um usuario.
      type: object
      properties:
        publicPlace:
          type: string
          description: Novo nome da rua, avenida, praca, etc.
          example: Avenida Central
        number:
          type: string
          description: Novo numero do imovel.
          example: '999'
        complement:
          type: string
          description: Novo complemento do endereco (apartamento, bloco, etc.).
          example: Casa
        neighborhood:
          type: string
          description: Novo nome do bairro.
          example: Centro
        city:
          type: string
          description: Nova cidade.
          example: Rio de Janeiro
        state:
          type: string
          description: Nova sigla do estado (UF).
          example: RJ
        postalCode:
          type: string
          description: Novo Codigo de Enderecamento Postal (CEP).
          example: 98765-432
    AddressesResponse:
      title: Resposta de Endereco
      description: Estrutura da resposta contendo os detalhes de um endereco.
      type: object
      properties:
        id:
          format: uuid
          type: string
          description: O ID unico do endereco.
          example: b1a7d7a3-2f18-4c91-a10d-9e4a5c0f3b99
        publicPlace:
          type: string
          description: Nome da rua, avenida, praca, etc.
          example: Rua das Flores
        number:
          type: string
          description: Numero do imovel.
          example: '123'
        complement:
          type: string
          description: Complemento do endereco (apartamento, bloco, etc.).
          example: Apto 202
        neighborhood:
          type: string
          description: Nome do bairro.
          example: Jardim Primavera
        city:
          type: string
          description: Nome da cidade.
          example: Sao Paulo
        state:
          type: string
          description: Sigla do estado (UF).
          example: SP
        postalCode:
          type: string
          description: Codigo de Enderecamento Postal (CEP).
          example: 01234-567
        isActive:
          type: boolean
          description: Indica se o endereco esta ativo.
          example: true
        createdAt:
          type: string
          format: date-time
          description: Data e hora de criacao do endereco.
          example: '2025-01-15T14:30:00-03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da ultima atualizacao do endereco.
          example: '2025-01-20T10:15:30-03:00'
...
    MenuItemResponse:
      title: Resposta de Item do Cardapio
      type: object
      description: Estrutura da resposta contendo os detalhes de um item do cardapio.
      properties:
        id:
          type: string
          format: uuid
          description: ID unico do item no sistema.
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Nome do item.
          example: Moqueca de Camarao
        description:
          type: string
          description: Descricao detalhada do item.
          example: Deliciosa moqueca de camarao com arroz e pirao.
        price:
          type: number
          format: double
          description: Preco do item em reais.
          example: 59.90
        availableForInStoreOnly:
          type: boolean
          description: Indica se o item so esta disponivel para consumo no local.
          example: false
        photoUrl:
          type: string
          description: URL da imagem do item no servico de armazenamento.
          example: https://storage.example.com/menu-items/123e4567-e89b-12d3-a456-426614174000.jpg
        restaurantId:
          type: string
          format: uuid
          description: ID do restaurante a que este item pertence.
          example: 123e4567-e89b-12d3-a456-426614174000
        isActive:
          type: boolean
          description: Indica se o item do cardapio esta ativo.
          example: true
        createdAt:
          type: string
          format: date-time
          description: Data e hora de criacao do item.
          example: '2025-01-15T14:30:00-03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da ultima atualizacao do item.
          example: '2025-01-20T10:15:30-03:00'
...